Create implied_vol_comments0912
This is the implied_vol function with added lines of code to identify if conditions are unfavourable. This is based on John's email and some things mentioned in the paper 'By Implementation' he recommended. All this code does is identify problem, there still needs to be added a way to solve these problems.